fix: move notificationsprovider
